[hwajun, gahee] feat : 쇼츠 통합 완료

+++ src/main/java/com/ohgoodteam/ohgoodpay/shorts/repository/search/ShortsSearchRepositoryCustom.java
@@ -1,5 +1,6 @@
 package com.ohgoodteam.ohgoodpay.shorts.repository.search;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import com.ohgoodteam.ohgoodpay.shorts.dto.response.search.ShortsSearchResponseDto.ShortsSearchResponse;
@@ -7,7 +8,6 @@ import com.ohgoodteam.ohgoodpay.shorts.dto.response.search.ShortsSearchResponseD
 public interface ShortsSearchRepositoryCustom {
     
     //검색 결과 조회 (가중치 로직, 점수 계산, 정렬, 페이징 처리)
-    
     List<ShortsSearchResponse> fetchExposure(
         String q,
         double wLike,
@@ -15,7 +15,7 @@ public interface ShortsSearchRepositoryCustom {
         double wHashtag,
         double wRecency,
         double tauHours,
-        Double lastScore,
+        BigDecimal lastScore,
         LocalDateTime lastDate,
         Long lastId,
         int limitPlusOne
