+++ src/main/java/com/ohgoodteam/ohgoodpay/shorts/service/search/ShortsSearchService.java
@@ -1,5 +1,6 @@
 package com.ohgoodteam.ohgoodpay.shorts.service.search;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import com.ohgoodteam.ohgoodpay.shorts.dto.response.search.ShortsSearchResponseDto.CursorResponse;
 
@@ -20,6 +21,6 @@ public interface ShortsSearchService {
             Integer limit,
             Long lastId,
             LocalDateTime lastDate,
-            Double lastScore
+            BigDecimal lastScore
     );
 }
